+++ qct-core/src/main/java/com/github/aleksikangas/qct/core/meta/ExtendedData.java
@@ -41,16 +41,21 @@ public record ExtendedData(String mapType,
   }
 
   public static final class Decoder {
-    public static ExtendedData decode(final FileChannel fileChannel, final long byteOffset) {
+    public static ExtendedData decode(final FileChannel fileChannel, final int byteOffset) {
       return new ExtendedData(QctReader.readStringFromPointer(fileChannel, byteOffset),
                               DatumShift.Decoder.decode(fileChannel,
-                                                        QctReader.readPointer(fileChannel, byteOffset + 0x04L)),
-                              QctReader.readStringFromPointer(fileChannel, byteOffset + 0x08L),
+                                                        QctReader.readPointer(fileChannel,
+                                                                              Math.toIntExact(byteOffset + 0x04L))),
+                              QctReader.readStringFromPointer(fileChannel, Math.toIntExact(byteOffset + 0x08L)),
                               LicenseInformation.Decoder.decode(fileChannel,
-                                                                QctReader.readPointer(fileChannel, byteOffset + 0x14L)),
-                              QctReader.readStringFromPointer(fileChannel, byteOffset + 0x18L),
+                                                                QctReader.readPointer(fileChannel,
+                                                                                      Math.toIntExact(byteOffset +
+                                                                                                      0x14L))),
+                              QctReader.readStringFromPointer(fileChannel, Math.toIntExact(byteOffset + 0x18L)),
                               DigitalMapShop.Decoder.decode(fileChannel,
-                                                            QctReader.readPointer(fileChannel, byteOffset + 0x1CL)));
+                                                            QctReader.readPointer(fileChannel,
+                                                                                  Math.toIntExact(byteOffset +
+                                                                                                  0x1CL))));
     }
 
     private Decoder() {
